+++ models.py
@@ -3,9 +3,11 @@ author:Sanidhya Mangal
 github:sanidhyamangal
 """
 
-from typing import Tuple
+from typing import Any, Dict, Tuple, Union
 
 import tensorflow as tf  # for deep learning stuff
+from tensorflow.keras.models import Model
+from tensorflow.keras.layers import Layer
 
 
 class BaseNetModel(tf.keras.models.Model):
@@ -20,6 +22,7 @@ class BaseNetModel(tf.keras.models.Model):
                  num_classes: int,
                  fine_tune_at: int = 0,
                  train_from_scratch: bool = False,
+                 custom_input_preprocessing:bool = False,
                  *args,
                  **kwargs):
         super().__init__(*args, **kwargs)
@@ -43,9 +46,16 @@ class BaseNetModel(tf.keras.models.Model):
         # # define preprocessing model right before input to the base model
         # self.preprocess = tf.keras.models.Sequential(_pre_model_arch)
 
+        self.custom_input_processing = custom_input_preprocessing
+        _img_shape = (224,224,img_shape[-1]) if self.custom_input_processing else img_shape
+
+        if custom_input_preprocessing:
+            self.preprocess_input = tf.keras.layers.Conv2D(filters=3, kernel_size=(2,3), strides=(1,2), input_shape=img_shape)
+        self.preprocess_input = self.model_config.get('preprocess_input')
+
         # define base model for the training purpose
         self.base_model = self.model_config.get('model_layer')(
-            input_shape=img_shape, include_top=False, weights='imagenet')
+            input_shape=_img_shape, include_top=False, weights='imagenet')
 
         if not train_from_scratch:
             # freeze the base models to restrict further training
@@ -69,8 +79,7 @@ class BaseNetModel(tf.keras.models.Model):
             tf.keras.layers.Dropout(kwargs.get('dropout', 0.2)),
             tf.keras.layers.Dense(num_classes)
         ])
-
-        self.preprocess_input = self.model_config.get('preprocess_input')
+        
 
     def _check_config_hyperparams(self,
                                   param_name: str,
@@ -91,6 +100,9 @@ class BaseNetModel(tf.keras.models.Model):
     def call(self, inputs, training):
         # call to the preprocessing input unit for the model nased preprocessing
         x = self.preprocess_input(inputs)
+
+        if self.custom_input_processing:
+            x = tf.image.resize(x, size=(224,224))
 
         # preprocessing layer called for data augmentation part
         # x = self.preprocess(x, training=training)